grsec: Apply grsecurity-2.9.1-3.2.20-201206182053.patch
commit 3c286bc9765b9df15ca80013c6dc5587852c2d47
Merge: d240764 3c05eb4
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:21:29 2012 -0400

    Merge branch 'pax-stable2' into grsec-stable2

commit 3c05eb4bdf944054112476abb781b016e5a36058
Author: Brad Spengler <spender@grsecurity.net>
Date:   Mon Jun 18 20:20:48 2012 -0400

    Update to pax-linux-3.2.20-test54.patch

 arch/parisc/include/asm/uaccess.h |    4 ++--
 arch/s390/include/asm/uaccess.h   |    4 ++--
 arch/tile/include/asm/uaccess.h   |    4 ++--
 arch/x86/include/asm/uaccess_32.h |    8 ++++----
 arch/x86/include/asm/uaccess_64.h |    8 ++++----
 tools/gcc/size_overflow_plugin.c  |    2 +-
 6 files changed, 15 insertions(+), 15 deletions(-)

Signature-tree: 9e723212ca78313cd92c41b8e20af5553dfc4faa

+++ arch/parisc/include/asm/uaccess.h
@@ -253,10 +253,10 @@ static inline unsigned long __must_check copy_from_user(void *to,
                                           const void __user *from,
                                           unsigned long n)
 {
-        int sz = __compiletime_object_size(to);
+        size_t sz = __compiletime_object_size(to);
         int ret = -EFAULT;
 
-        if (likely(sz == -1 || !__builtin_constant_p(n) || sz >= n))
+        if (likely(sz == (size_t)-1 || !__builtin_constant_p(n) || sz >= n))
                 ret = __copy_from_user(to, from, n);
         else
                 copy_from_user_overflow();
